Added total sacks by quarter, total yards minus penalties
The included code was created in an effort to better identify the number of sacks over the entire season, then identify
the quarters that the teams are most at risk for being sacked. The other code is completed in effort to identify
the number of yards gained per team over the entire season minus the yards lost due to penalties.

The next steps I believe will be to identify which quarters each team is most at risk for penalties, and the average
amount of yards given up per quarter as a result of penalties.